argoslabs/ocr/abbyy/requirements.txt
# pip dependent packages

# default for alabs
alabs.common

#for ABBYY
python-magic-bin==0.4.14
